+++ README.md
@@ -6,7 +6,10 @@ To learn database system.
 ## What is makedb?
 makedb is a persisent k-v database that based on bitcask.bitcask's paper is [here](https://riak.com/assets/bitcask-intro.pdf). 
 
-## Usage
-```go
-to-do
-```
+## Task
+- [x] Get/Put k/v with string.
+- [ ] Support http protocol.
+- [ ] Support redis protocol.
+- [ ] Rotate active file.
+- [ ] Write ahead log.
+- [ ] Get/Put k/v with list.
